# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: PSCNC · API de Firma Electrónica No Cualificada
  description: 'Contrato público de FirmasNoCualificadas.


    Reglas que un integrador debe conocer antes de programar:


    - La decisión de identidad llega **tomada por el tenant**; el servicio la asienta como evidencia y
    no la revisa.

    - El OTP verificado por el tenant viaja como prueba, nunca el código.

    - El modo predeterminado es *hash-only*: el documento no se envía.

    - Las escrituras exigen `Idempotency-Key`; una confirmación repetida devuelve el acta original.

    - Todo rechazo trae un `motivo` de un enumerado estable, nunca un mensaje libre.'
  version: 0.1.0
paths:
...
  /v1/signing-sessions:
    post:
      tags:
      - firma
      summary: Crear Sesion
      description: Inicia una sesión de firma vinculada a un onboarding aprobado.
      operationId: crear_sesion_v1_signing_sessions_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_crear_sesion_v1_signing_sessions_post'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SigningSessionCreated'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/signing-sessions/{session_id}/confirm:
    post:
      tags:
      - firma
      summary: Confirmar Firma
      description: Verifica el consentimiento y aplica la firma PAdES-B-T con sello de tiempo.
      operationId: confirmar_firma_v1_signing_sessions__session_id__confirm_post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmSignatureRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SigningSessionCompleted'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/signing-sessions/{session_id}/evidence:
    get:
      tags:
      - evidencia
      summary: Obtener Evidencia
      description: Entrega el expediente de evidencias y el documento firmado (URLs temporales).
      operationId: obtener_evidencia_v1_signing_sessions__session_id__evidence_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvidenceBundle'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    Body_crear_sesion_v1_signing_sessions_post:
      properties:
        onboarding_token:
          type: string
          title: Onboarding Token
        pdf_document:
          type: string
          contentMediaType: application/octet-stream
          title: Pdf Document
        metadata:
          anyOf:
          - type: string
          - type: 'null'
          title: Metadata
      type: object
      required:
      - onboarding_token
      - pdf_document
      title: Body_crear_sesion_v1_signing_sessions_post
    ConfirmSignatureRequest:
      properties:
        consent_otp_code:
          type: string
          maxLength: 10
          minLength: 4
          pattern: ^[0-9]+$
          title: Consent Otp Code
        consent_statement:
          type: string
          maxLength: 4000
          minLength: 1
          title: Consent Statement
        visual_signature_enabled:
          type: boolean
          title: Visual Signature Enabled
          default: true
        signature_page:
          type: integer
          minimum: 1.0
          title: Signature Page
          default: 1
        signature_coordinate_x:
          type: number
          minimum: 0.0
          title: Signature Coordinate X
          default: 100.0
        signature_coordinate_y:
          type: number
          minimum: 0.0
          title: Signature Coordinate Y
          default: 150.0
        signature_width:
          type: number
          exclusiveMinimum: 0.0
          title: Signature Width
          default: 180.0
        signature_height:
          type: number
          exclusiveMinimum: 0.0
          title: Signature Height
          default: 60.0
      additionalProperties: false
      type: object
      required:
      - consent_otp_code
      - consent_statement
      title: ConfirmSignatureRequest
      description: 'Cuerpo de `POST /v1/signing-sessions/{id}/confirm`.


        El código OTP se recibe para su verificación pero nunca se persiste ni se

        registra: en la evidencia solo queda su hash SHA-256.'
...
    EvidenceBundle:
      properties:
        signing_session_id:
          type: string
          title: Signing Session Id
        status:
          $ref: '#/components/schemas/SigningStatus'
        signed_document_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Signed Document Url
        evidence_report_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Evidence Report Url
        url_expires_in_seconds:
          type: integer
          title: Url Expires In Seconds
        original_document_hash:
          type: string
          title: Original Document Hash
        signed_document_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Signed Document Hash
        verifications:
          $ref: '#/components/schemas/VerificationSummary'
      additionalProperties: false
      type: object
      required:
      - signing_session_id
      - status
      - url_expires_in_seconds
      - original_document_hash
      - verifications
      title: EvidenceBundle
      description: 'Respuesta de `GET /v1/signing-sessions/{id}/evidence`.


        Las URLs son pre-firmadas de S3 con vigencia limitada (300 s por defecto).'
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SigningSessionCompleted:
      properties:
        signing_session_id:
          type: string
          title: Signing Session Id
        status:
          $ref: '#/components/schemas/SigningStatus'
        signed_document_hash:
          type: string
          title: Signed Document Hash
        user_certificate_serial:
          type: string
          title: User Certificate Serial
        signature_format:
          type: string
          title: Signature Format
        timestamp:
          $ref: '#/components/schemas/TimestampInfo'
      additionalProperties: false
      type: object
      required:
      - signing_session_id
      - status
      - signed_document_hash
      - user_certificate_serial
      - signature_format
      - timestamp
      title: SigningSessionCompleted
      description: Respuesta de la confirmación de firma.
    SigningSessionCreated:
      properties:
        signing_session_id:
          type: string
          title: Signing Session Id
        status:
          $ref: '#/components/schemas/SigningStatus'
        original_document_hash:
          type: string
          title: Original Document Hash
        created_at:
          type: string
          format: date-time
          title: Created At
        expires_at:
          type: string
          format: date-time
          title: Expires At
      additionalProperties: false
      type: object
      required:
      - signing_session_id
      - status
      - original_document_hash
      - created_at
      - expires_at
      title: SigningSessionCreated
      description: Respuesta de `POST /v1/signing-sessions`.
    SigningStatus:
      type: string
      enum:
      - INITIALIZED
      - ONBOARDING_COMPLETED
      - SIGNING_COMPLETED
      - FAILED
      - REVOKED
      - COMPROMISED
      title: SigningStatus
      description: Estados de la máquina de estados de la sesión de firma.
    TimestampInfo:
      properties:
        authority:
          type: string
          title: Authority
        serial:
          anyOf:
          - type: string
          - type: 'null'
          title: Serial
        time:
          type: string
          format: date-time
          title: Time
      additionalProperties: false
      type: object
      required:
      - authority
      - time
      title: TimestampInfo
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VerificationSummary:
      properties:
        aml_pep_checked:
          type: boolean
          title: Aml Pep Checked
        biometric_score:
          type: number
          title: Biometric Score
        liveness_detected:
          type: boolean
          title: Liveness Detected
        identity_match_approved:
          type: boolean
          title: Identity Match Approved
      additionalProperties: false
      type: object
      required:
      - aml_pep_checked
      - biometric_score
      - liveness_detected
      - identity_match_approved
      title: VerificationSummary